refactor: 회원가입 입력 폼 유형별 test 파일 분리
